Added test data
Test data was added, commented code for training/testing split. The testing data is currently not cleaned, but based on predictor selection, it appears that increasing the data size improved accuracy (70% to 73%). Also, the K value now changes to a much larger number. Would be a good idea to rerun cross-validation with more k values. It is computationally costly, current K values (up to 35) takes about 250 seconds to run all the code (~ 4 min)